@@ -1,19 +1,17 @@
 ///
 //  DEPENDENCIES
 ///
-
 import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { HttpClientModule } from '@angular/common/http';
-import { HttpModule, JsonpModule } from '@angular/http';
-
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { ChartsModule } from 'ng2-charts';
+import { Ng2OrderModule } from 'ng2-order-pipe';
+import { FormsModule } from '@angular/forms';
 
 ///
 //  COMPONENTS
 ///
-
 import { AppComponent } from './app.component';
 import { AssociateListComponent } from './components/associate-list/associate-list.component';
 import { BatchListComponent } from './components/batch-list/batch-list.component';
@@ -21,27 +19,46 @@ import { ClientMappedComponent } from './components/client-mapped/client-mapped.
 import { CreateUserComponent } from './components/create-user/create-user.component';
 import { LoginComponent } from './components/login/login.component';
 import { ClientListComponent } from './components/client-list/client-list.component';
+import { FormComponent } from './components/form-component/form.component';
 import { NavbarComponent } from './components/navbar/navbar.component';
 import { HomeComponent } from './components/home/home.component';
+import { RootComponent } from './components/root/root.component';
+import { SkillsetComponent } from './components/skillset/skillset.component';
+import { BatchDetailsComponent } from './components/batch-details/batch-details.component';
+import { AssociateViewComponent } from './components/associate-view/associate-view.component';
 
 ///
 //  SERVICES
 ///
-import { RequestService } from './services/request.service';
-
+import { RequestService } from './services/request-service/request.service';
 import { AssociateService } from './services/associates-service/associates-service';
-import { ClientService } from './services/clients-service/clients-service';
-import { ClientMappedService } from './services/client-mapped-service/client-mapped-service.service';
-import { AuthenticationService } from './services/authentication/authentication.service';
+import { ClientListService } from './services/client-list-service/client-list.service';
+import { ClientMappedService } from './services/client-mapped-service/client-mapped-service';
+import { AuthenticationService } from './services/authentication-service/authentication.service';
+import { SearchFilterPipe } from './pipes/search-filter/search-filter.pipe';
+import { BatchService } from './services/batch-service/batch.service';
+import { SkillsetService } from './services/skill-set-service/skill-set.service';
+import { DataSyncService } from './services/datasync-service/data-sync.service';
+import { CreateUserService } from './services/create-user-service/create-user.service';
+
+///
+//  FILTERS
+///
+///
+
+import { AssociateSearchByTextFilter } from './pipes/associate-search-by-text-filter/associate-search-by-text-filter.pipes';
+
+///
+//  SECURITY
+///
+import { JwtInterceptor } from './interceptors/jwt.interceptor';
 
 ///
 //  CONSTANTS
 ///
-
 import { appRoutes } from './routing/routes';
-import { RootComponent } from './components/root/root.component';
-
-
+import { RouterLinkStubDirective, RouterOutletStubComponent } from './testing-helpers/router-stubs';
+import { FooterComponent } from './components/footer/footer/footer.component';
 
 
 @NgModule({
@@ -50,22 +67,46 @@ import { RootComponent } from './components/root/root.component';
     NavbarComponent,
     HomeComponent,
     AssociateListComponent,
+    AssociateSearchByTextFilter,
     BatchListComponent,
     ClientMappedComponent,
+    FormComponent,
     ClientListComponent,
     LoginComponent,
     CreateUserComponent,
-    RootComponent
+    SearchFilterPipe,
+    BatchDetailsComponent,
+    RootComponent,
+    SkillsetComponent,
+    AssociateViewComponent,
+    RouterLinkStubDirective,
+    RouterOutletStubComponent,
+    FooterComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpClientModule,
-    FormsModule,
-    HttpModule,
-    RouterModule.forRoot(appRoutes)
+    RouterModule.forRoot(appRoutes),
+    ChartsModule,
+    Ng2OrderModule
   ],
-  providers: [AssociateService, ClientService, ClientMappedService, AuthenticationService, RequestService],
+  providers: [
+    AssociateService,
+    ClientListService,
+    ClientMappedService,
+    AuthenticationService,
+    RequestService,
+    SkillsetService,
+    BatchService,
+    CreateUserService,
+    SkillsetService,
+    DataSyncService,
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: JwtInterceptor,
+      multi: true
+    }],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
