@@ -3,18 +3,18 @@
  * @description Receives user inputs from form and submits them to the back-end for validation
  */
 import { Component, OnInit } from '@angular/core';
-import { environment } from '../../../environments/environment';
-import { AuthenticationService } from '../../services/authentication/authentication.service';
+import { AuthenticationService } from '../../services/authentication-service/authentication.service';
 import { Router } from '@angular/router';
+import { AutoUnsubscribe } from '../../decorators/auto-unsubscribe.decorator';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.css']
 })
+@AutoUnsubscribe
 export class LoginComponent implements OnInit {
   //url to REST endpoint
-  public loginUrl:string = environment.url + 'TrackForce/track/user/submit';
   public username: string;
   public password: string;
 
@@ -38,7 +38,14 @@ export class LoginComponent implements OnInit {
   *
   */
   ngOnInit() {
-    this.authService.logout();
+    let user = this.authService.getUser();
+    if (user != null){
+      if(user.tfRoleId === 4){
+        this.router.navigate(['associate-view', user.userId]);
+      }
+      this.router.navigate(['root']);
+    }
+
   }
 
   /**
@@ -49,15 +56,20 @@ export class LoginComponent implements OnInit {
   *@param none
   *
   */
-  login(){
-    this.authService.login(this.username, this.password, this.loginUrl).subscribe(
+  login() {
+    this.authService.login(this.username, this.password).subscribe(
       data => {
-        console.log(data)
-        //navigate to home page if return is valid
-        this.router.navigate(['home']);
+        debugger;
+        let user = this.authService.getUser();
+        //navigate to appropriate page if return is valid
+        if(user.tfRoleId === 4){
+          this.router.navigate(['associate-view', user.userId]);
+        } else {
+          this.router.navigate(['root']);
+        }
       },
       err => {
-        console.log(err + " Error Occurred");
+        debugger;
         this.authService.logout();
       }
     );
