@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  url: "http://52.201.234.9:8085/"
 };
