@@ -2,10 +2,16 @@ package com.revature.model;
 
 import java.math.BigDecimal;
 
-//This class is the Java representation of a JSON bean to be set to Angular
-//Jackson will automatically convert the Java Bean to JSON
-//This beans contains user roleId, username, JWT token
-//Produced by the REST service
+/**
+ * 
+ * @author Michael Tseng
+ * 
+ * This class is the Java representation of a JSON bean to be set to Angular
+ * Jackson will automatically convert the Java Bean to JSON
+ * This beans contains user roleId, username, JWT token
+ * Produced by the REST service
+ *
+ */
 public class UserJSON {
 	
 	private String username;
@@ -13,6 +19,8 @@ public class UserJSON {
 	private BigDecimal tfRoleId;
 	//token is the JWT token generated
 	private String token;
+	//user id
+	private int userId;
 	
 	public UserJSON() {
 		super();
@@ -42,9 +50,18 @@ public class UserJSON {
 		this.token = token;
 	}
 
+	public int getUserId() {
+		return userId;
+	}
+
+	public void setUserId(int userId) {
+		this.userId = userId;
+	}
+
 	@Override
 	public String toString() {
-		return "UserJSON [username=" + username + ", tfRoleId=" + tfRoleId + ", token=" + token + "]";
+		return "UserJSON [username=" + username + ", tfRoleId=" + tfRoleId + ", token=" + token + ", userId=" + userId
+				+ "]";
 	}
 
 
